微変更
サブウインドウがテーマ変更されたときに更新されるようにしたい
